added `to_sparse` in heat/sparse/manipulations.py
